requirements.txt
# Combined requirements for all tasks

# LangChain
langchain>=0.3.0
langchain-core>=0.3.0
langchain-community>=0.3.0
langchain-openai>=0.2.0
langchain-chroma>=0.1.0
langchain-text-splitters>=0.3.0

# Vector Store
chromadb>=0.5.0

# LLM APIs
openai>=1.3.0
google-generativeai

# Document Loaders
pypdf>=4.0.0
unstructured>=0.10.0
markdown>=3.5.0

# Environment & Utilities
python-dotenv>=1.0.0
requests>=2.31.0
beautifulsoup4>=4.12.0
tqdm>=4.66.0
typing-extensions>=4.5.0
pydantic>=2.0.0
